+++ packages/graphiql-react/src/editor/header-editor.tsx
